backend/requirements.txt
openai>=1.3.5           # Required for OpenAI SDK (make sure it's >=1.3 for assistants support)
python-dotenv           # To load environment variables from .env
openai-agents[litellm]
configparser
openai-agents
anthropic
google-generativeai>=0.3.1
google-cloud-aiplatform
requests
pillow==9.5.0
google-search-results
openai-whisper
moviepy==1.0.3
fastapi[standard]